[ADD] Right Sidebar + CSS + FONT
