+++ src/components/RightSidebar/RightSidebar.css
@@ -1,3 +1,4 @@
 .right-sidebar {
   padding: 2em;
+  text-align: left;
 }
